fix(terminology): moved more class defs
Part of ENG-330
